# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/techhunt-login": {
      "post": {
        "operationId": "AuthController_techhuntLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechhuntLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TechHunt login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login through TechHunt API",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          }
        },
        "summary": "Refresh access token using refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log out and revoke refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All sessions logged out successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Log out from all devices by revoking all refresh tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/signup": {
      "post": {
        "operationId": "AuthController_adminSignup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin account created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Admin already exists"
          }
        },
        "summary": "Register a new admin account",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/signin": {
      "post": {
        "operationId": "AuthController_adminSignin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSigninDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin signed in successfully"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Sign in as admin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/admin/profile": {
      "get": {
        "operationId": "AuthController_getAdminProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Admin profile retrieved"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Admin access required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get admin profile (admin only)",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "DEPA API",
    "description": "The DEPA API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TechhuntLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username or email address",
            "example": "mathuros@playpark.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "TH@digital"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AdminSignupDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Admin username",
            "example": "admin123"
          },
          "email": {
            "type": "string",
            "description": "Admin email address",
            "example": "admin@depa.com"
          },
          "password": {
            "type": "string",
            "description": "Admin password",
            "example": "SecureAdminPass123!"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "AdminSigninDto": {
        "type": "object",
        "properties": {
          "usernameOrEmail": {
            "type": "string",
            "description": "Admin username or email address",
            "example": "admin123"
          },
          "password": {
            "type": "string",
            "description": "Admin password",
            "example": "SecureAdminPass123!"
          }
        },
        "required": [
          "usernameOrEmail",
          "password"
        ]
      },
...
    }
  }
}